Refactor: move the login logic to the controller
